Tweak camera FOV based on smovespeed
Basically a lerp between two FOV floats over a duration. Current FOV and Target FOV. Target FOV is clamped to ensure it never goes above or below hard limits. It's modified by newSpeed which is publicly called from the Change Chunk Move Speed method in the LevelGenerator, so that the FOV changes at the same time we're altering the chunk move speed for greater visual effect.